requirements.txt
asgiref==3.7.2
cloudinary==1.33.0
dj-database-url==2.0.0
dj3-cloudinary-storage==0.0.6
Django==3.2.20
django-cors-headers==4.2.0
django-shortcuts==1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
gunicorn==21.1.0
inflection==0.5.1
Pillow==10.0.0
psycopg2==2.9.6
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.16
